fitnet-backend/requirements.txt
django
djangorestframework
geopy
drf_yasg